Revamped Styles internals; fixed fallback between states.
When seeking a style for a non-default state, we now immedaiately fall back to
the default state if one is not found. Then we proceed to our containing
group's stylesheet, where we continue to fall back as early as possible to the
default style if a binding for the non-default style cannot be found.

The reasoning is clear from example: consider that one configures the font for
a button either on the button, or in the group that contains the button. One
would expect that font to be used for both the DOWN and DEFAULT state,
regardless of whether there happened to be a custom DOWN font configured in the
top-level style sheet.