Reinstated the constructors that take styles.
A Styles instance is like a CSS class, and in GWT one of the most useful things
we did was to create a Widgets class which had convenience methods for
constructing Buttons, Labels, etc. with a CSS class directly instead of having
to do:

  Button b = new Button("Click me!");
  b.addStyleClass("awesome");

The reason is because one sets custom styles on things a lot. Not always, mind
you, but often enough that it's annoying not to be able to pass the style name
in the constructor.

TPUI ameliorated this somewhat by having setStyles() be fluent, but it was
still too annoying to see 25 calls to setStyles in the creation of a UI.
Setting custom styles is central enough that I'd prefer the conciseness of
passing those styles in the constructor. I'd hate to force someone to create an
Elements class for creating TPUI elements, not least because it would conflict
with the parent class of Group!