Revamp of Input	based on Tim's suggestions.
There's no longer a distinction between a reactor and a bounded pointer
listener. A reactor now defines a hit test and expiry functions, and provides a
pointer listener to be notified when a pointer click triggers the reactor.
Subsequent drag and release events are dispatched to the triggered reactor's
listener.

Action is now an adapted Pointer.Listener that simply calls "onTrigger" when
the pointer is clicked and discards drag and release events.