Document Selector; decouple it from ClickableText
I guess this should stick around, though I'm not very happy with the
implementation. It feels like a lot of code for what it's doing, but it
does make a selectable list easier to implement.

I'm also unhappy with how it fiddles with the SELECTED flag directly,
which Element uses for other things. It works out currently, but that
feels fragile.

I'm also also unhappy with dispatching Element from selected. You almost
certainly have single subclass of Element in the elements you're
selecting on, so Selector should be genercized on that.  You're going to
need that specific class as Element itself isn't going to have any
interesting information exposed on it. I made Elements take a type
parameter of what it contained, but it made dealing with Root and Group
an exercise in frustrating wildcards and other generics bullshit.  For
now you just end up casting in whatever's using the selector to get the
specific information out.

Documenting it made me want to extract Clickable from
ClickableTextWidget, though I could be convinced it's better to avoid
Clickable until we have more clickable widgets in tripleplay itself.