Add region-based dispatching for mice
This abstracts out Input to just care about Events.Position, which made
for a pretty simple implementation of MouseInput with all the region
stuff.

It's possible now to do Input-like dispatching for Touch as well, but
it's a little less straightforward. How does it choose between multiple
touches matching multiple regions to initialize? Beyond that, I'm not
sure if there's the same demand for region-based dispatching for touch
events, so I didn't add it.