Clarified the comments and code. I'm being OCD, but I did fix a bug.
Having both _active and _lastMoved seemed weird, since there should only ever
be one target for mouse events, so why track that target in two variables. Now
we just have _target and we decide whether to update it on mouse move depending
on whether the mouse button is down or not.

The fixed bug is the rechecking of the target on mouse release, otherwise you
could press the mouse down on one region, keep it down and hover over to
another region, then release the mouse, and subsequent wheel events would go to
the old mouse downed region, not the region over which you're now hovering.

We might want to consider checking Mouse.ButtonEvent.getPreventDefault and not
setting _down if default has been prevented, so that an app could only "grab"
the mouse when the first mouse button was pressed (or middle, or whatever it
prefers), and thus other mouse buttons would not cause hovering to stop working
until the mouse was released. It's probably overkill, so we can do it later if
we need it.