Specify columns via a base instance a la Styles
This still requires a static import to be semi-readable, but I'm getting
more ok with those. I'm already at the point where I static import
Style.* to use tripleplay, and this feels similar.

I'm a little queasy about calling it COL and not COLUMN, but going down
to COL makes me feel ok with getting rid of the int constructor.