Stop pussyfooting around Value on settable items
This exposes the Value directly on Selector and Slider, which were
previously exposing a setter function and a ValueView. The Value lets
users updateForce and thereby cause an emit, but they could've done that
anyway with a couple calls to set. This just removes a bunch of
indirection and makes the reactive version central, in acceptance of the
comments at
https://github.com/threerings/tripleplay/commit/0d2f7768e44c5ae7c453e9d3d31a02a541e9f14.