Add a delegate to Root for uncaught pointer events
If none of the Elements of the Root are hit by the pointer event, the
delegate gets a chance to handle it. If it chooses to, the event doesn't
propagate to other Roots and drag and end events will be dispatched to
the delegate as well.

This allows for Roots to contain things like modal dialogs or popups
that want to block other clicks and dismiss when the user clicks outside
the root.