Added a general system of layout constraints.
Layout constraints can influence an element's preferred size, and in the case
of TextWidget and TextConstraint, can influence the widget's text size (which
is later augmented with an icon and insets and whatnot).

It is also possible for a layout manager to honor general constraints, but at
the moment, no such constraints exist as everything is handled through the
preferred size communication.

This also nixes setMaxText(tmpl) which is now
setConstraint(Constraints.fixSize(tmpl)) or minSize(tmpl).