Revamped cursor implementation.
The cursor is now a separate layer, which vastly reduces regeneration of the
text layer, and avoids funny artifacts when the cursor ends up on the right end
of the text layer, which is then not big enough to handle it.

I also modified the click to insertion point computation so that if you click
closer to the left side of a character, the cursor goes before the character
rather than after it. Thus begins the slippery slope of ten thousand changes to
make text entry not suck.