Improved click handling for overlapping roots.
Previously, one could click through one root and into another. This is weird.
Now if a "higher" root contains the mouse click, we stop searching even if it
declares that the pointer clicked nothing interesting.

We still need to determine higher based on depth rather than the order in which
the roots were added to the Interface.