Revamped event handling to use new standard PlayN per-layer event dispatch.
This means that a TPUI Pointer.Listener does not need to be registered. The
Root elements automatically register to listen for pointer events within their
bounds. This allows TPUI to coexist with game-specific pointer event handling
much more nicely.

A Keyboard.Listener still needs to be registered as PlayN has no composable way
to listen for key events. Probably we should provide a delegate mechanism like
we used to have for pointer events, so that keyboard events can be sent to the
delegate when there's no focused element.