Fold PopupTextEntry's functionality into Field.
I might add a mode down the road that tells it to force into popup mode in Java land for
functionality testing. Or now that I think about it, maybe the thing to do is to create a version
of the Java PlayN platform that acts as much like a device platform as it can, for testing. In this
case that would mean pretending it doesn't have a hardware keyboard.