Revamped how background insets are configured. Added Background.subImage().
Now instead of Background.blank(5) you use Background.blank().inset(5).

Apologies for the API disruption, but it was getting painful to have three
copies of every Background constructor method, and it was already verbose to
pass the insets up through every Background subclass constructor. Now insets
are unfortunately mutable, but we don't have sixty copies of the names top,
left, bottom and right in the code.