Revamp of TextWidget hierarchy, added CheckBox.
Having TextWidget expose a text and label property was a little wonky. For
example, Field thus inherited the ability to have an icon, which is crazy.

I revamped TextWidget to leave the maintenance of a text and/or label property
to its children. I changed said children to maintain their icon as a Value,
which they should have done all along, as it's less code and was clearly wanted
because we has iconSlot().

This also resulted in some constructor massaging so that you can pass in an
initial icon just as you pass in initial text. I removed the constructors that
take styles, because it's easy enough (and idiomatic) to write

  new Label("Foo").addStyles(...)

etc. Hopefully that won't break a bunch of callers. I did leave in a deprecated
setIcon() on clases that used to have it, but callers should supply an icon at
construct time or use icon.update(). Anyone who was using iconSlot() should now
use icon.slot().

The CheckBox leverages this cleaned up world order, to provide a simple
checkbox that can either use a check character (one is provided by default) or
a check image. It does not also display text (like Java's JCheckBox) because I
wanted to be able to set the background for a checkbox and have that go just
below the check. To keep that property but still have text would have required
great hackery, which didn't seem worth it. It's easy enough to write:

  new Group(AxisLayout.horizontal()).add(new CheckBox(), new Label("Foo"))

when you want a checkbox next to some text. In fact, I think I'll add some
convenience methods to AxisLayout that will make that even easier.