Don't try to be too fancy with our type parameter.
If we demand a type <T extends Element<T>> then Java will reject a concrete
type CustomGroup extends Group (which extends Elements<Group>). We don't need
the type to conform in such a detailed manner, we can just request something
that's a subtype of Element and be on our way.