Revamp of transition between animations.
When an animation completes, we want to process as many of the next animations
as possible on the same frame. For example, if a tween completes, and is
followed by three actions, those actions can all be processed immediately on
the same frame, we don't need to wait one frame for each action.

Additionally, make Action animations not consume any time, thus if they are
initialized with some left over time from the previous animation, they'll
continue to return that time as left over, so that they consume zero time. Then
if a new tween starts following the previous tween, it will be positioned as if
it started precisely when the previous tween ended.

This smooths things out when one has a complex string of animations that
involve adding and removing layers and chaining multiple tweens.