Removed the abomination that was text editing and focus management.
All Fields now use the text editing popup, so that text editing can be handled
by the native operating system where cut and paste, and foreign language input,
and text selection/editing and cursor navigation keyboard shortcuts work and
behave as the user expects.

If you are foolish enough to believe that you can reimplement the hundreds of
thousands of lines of code that go into even the least robust text editing
framework, you'll have to do it in some other library.