Revamped page flip transition using a shader.
Now we have real 3D perspective effects. Whee!

Also fixed an issue where the slide transition left the old screen with a wonky
translation.

At some point I'll generalize the RotateYShader to allow rotation in X or Y,
and I'll add the flip over transition so popular on iOS (which just requires a
rotation with the eye in the center of the page and a rotation of the new
screen into place at the right time).