Properly prepare the particle transform.
We concatenate all layer transforms up to the emitter layer, and we include the
root transform as well to get the appropriate device scale and rotation.

There is still an issue when device rotation is baked into the root transform:
particle velocities don't automatically account for the root transform, so
"down" will be different depending on the device orientation. When I have more
time to twiddle things, I'll make Gravity account for the device orientation.