Always clear out the cached preferred size when an invalidate is requested.
There are circumstances when a preferred size is requested multiple times with different hints
between validations (for instance, if a Root is packed, then a child size changes and it is
packed again), and the second preferred size isn't calculated because the setSize() requested an
invalidation but since it was already invalid, the preferred size wasn't cleared out.

If this is too weird, my current problem could probably be fixed by forcing a validation of the Root
in a specific spot in what I'm doing, but we've seen problems with this in other cases too, and I
think this is probably the right solution there.

I'm still learning my way around tripleplay though, so RFC 'n stuff.