Oops, previous commit contained new Sync DB bits.
Sync DB is a framework for maintaining a persistent storage on the client
(key/value and key/set) which is synced (via a server) across a collection of
devices. With appropriate hooks, the server can also extract data from syncs
that are associated with moves in multiplayer games and inject modifications
into the opponent's synced database to propagate moves to the opponent's
client(s).