Do not require the pointer to be inside our bounds to update the slider.
That results in super annoying behavior. If you've pressed down on the slider
thumb, you can wang your finger all over the place and we are going to update
the slider to reflect your wishes. We're not going to demand that you keep your
finger in the tiny little trough that is the slider's visualization and refuse
to acknowledge any updates outside of that realm.