Clear the layout data if the size != preferred size
This was just bailing out too early. During size calculation, an element such as a label with
TEXT_WRAP generates layout data that relies on the hints. But later, during layout, the size could
easily be 1) different from the hint and 2) the same as the last known size. Such elements will
then be stuck with their old layout data.

I put in an issue for this for posterity:

   https://github.com/threerings/tripleplay/issues/19