Allow one to call preferredSize on a root.
This is useful if you need to do something more complex than is possible via
the pack methods. In my case, I need to compute the preferred size of a screen,
given a width constraint, and then size the screen to the larger of the device
screen height and the screen's preferred size (in the latter case, one will
scroll up and down to see the whole screen).