Expose the Element selection state via an interface
It is perhaps odd that Togglable extends Clickable, but it makes sense for the use cases and
alternative names like ClickableSelectable seem unwarranted.

The need for this was made more urgent recently because Selector no longer works with buttons
(0846e9f). This probably makes sense. After some experimentation, I found they just fight over the
selection state in one way or another. Furthermore, it's not possible to get Button or Label to
work without Selector because setSelected is package private (which also probably makes sense).
So now we provide an explicit structure for exposing the selected state. Selector deals only with
that. So now other element types can participate with minimal hassle (and other Selector-like
classes can be written outside the ui package).