Make getStyleClass abstract.
Having a default implementation that returns getClass() is dangerous because it then becomes very
easy to mysteriously lose styles if a new subclass is created.

Also make SizableWidget abstract because it is generic and Shim can be used for every case I found
that was using a direct SizableWidget instance. Shim should maybe be changed to extend SizableWidget
so we get access to the resizing, but I suppose there's something to be said for keeping it as
simple as possible.