By default, always show the native field.
Even if we get text rendering to perfectly match what the native field on iOS does, there will still
probably be a small visual hiccup when the native field is displayed on top of the TP field. We
still need the option to only show the TP field sometimes though, for cases where we expect dialogs
to pop up on top of some layout that contains fields, and in cases where we maybe don't want a ton
of native fields on the screen, such as a table with editable cells.