Only set_Text() if it would be a value change.
In a secure text field in iOS, they have a nifty default feature where it shows you the last
character printed for a brief period, to help avoid mistyping a password. However, if you set_Text()
on a secure text field, it always shows all dots, even for the last character. So, to preserver
the secure text entry feature, we need to avoid setting the value of the native field when we can.