Cope with modifications made to properties while we're syncing.
We keep a modification count for each property, and take a snapshot of that
count when we start a sync. When the sync completes, we only clear out the
modified flag for properties in the original snapshot, and for which the
modification count has not increased.

We could just compare the values that were synced, but that would mean grinding
through the database to read all those values on sync complete which is more
expensive than I'd like. If you change a property and then change it back while
a sync is going on, you're just going to sync that property again next time.