Implemented bouncing-at-the-ends scrolling like iOS does.
I think this does a much better job of emulating iOS than iScroll or any of the
other JavaScript implementations I've seen (I do things based on actual
physical simulation, mostly, unlike most of those).

I also lowered the default friction a smidge, as the default seemed a bit
tighter than iOS does it.

Lastly, I switched back to absolute event coordinates rather than layer local.
I'm probably going to find out why I made that switch in the first place (since
I didn't document why at the time), but doing things layer local means that if
you make the Flicker a pointer listener on the layer that's going to be
scrolled (like a group layer, which is a natural thing to do), then all hell
breaks loose because the mouse positions are adjusted by the scrolling as you
drag and everything is wonky.

I also also added a flicker demo. It's not pretty but it shows off the goods.