Added clear() for nuking any lingering animations.
I use this in my screen stack when I know that I've transitioned away from a
screen that will be totally recreated when it is again shown. This ensures that
no lingering animations hang on to references that prevent big swathes of
memory from being collected while that screen lays quiescent on the stack.