Reuse our background if our size didn't change.
This obviates a massive heap of background recreation. Previously every time a
button was clicked all the backgrounds between it and the root got destroyed
and recreated. This isn't always expensive, but in the case of, say, a
round-rect background, it would redraw the round rect image.

And in the case of a project-local "draw my image into a surface and let the
image get garbage collected" background that I wrote, it's essential not to
reload the image every time anything in the UI is invalidated.