Keep a copy of our size, not the actual reference.
Otherwise we might be passed the _size field from and Element, and when we
later go to check whether the element's changed size is different from the
background's instantiated size, it won't be, even though _size changed, because
the background is hanging onto _size.

Reading IDimension is safe, retaining it, not necessarily so.