Introducing Menu
Incubated in tripleshim, provides a small framework for incorporating menus within the game. A
Menu is basically an animated widget with some item and triggering logic. MenuHost manages a
currently active menu at whatever layer you tell it is the "root".

Documentation and demo screen should be pretty self explanatory.