Fix menu item handling
This moves all the pointer tracking into the menu instead of partially and incorrectly handling
pointer events in MenuItem. So now if a menu is display and an item is clicked, it highlights on
pointer down. Dragging causes other items to highlight. If the pointer is then released over an
item, it triggers. If not, nothing happens. This creates a small problem for ShowText.WHEN_ACTIVE
menus, but I think it's solvable.