Demonstrate menus that kick off from onPointerStart
This results in the best user experience, we just need something besides Button to serve as the
trigger. ComboBox is a natural choice, but that confuses things by also having a Field.

The demo highlights the issue with the WHEN_ACTIVE text mode. It's supposed to expand and show the
text when selected but for some reason isn't relaying out. I'll have to read the original
RadialMenu that was using it to see how it worked.