Debounce Button clicks for 500ms.
You almost never want a button to be pressed twice in rapid succession, and if
you do, you can configure a custom style on it. I was tempted to make it longer
than 500ms, but that should hopefully be long enough to catch standard wacky
user behavior where they double click everything.