Added support for cuddling icon up next to text.
Normally the icon is always laid out in the left-most (or top-most, etc.)
position and the label is aligned in the remaining space. This allows one to
instead have the icon always be next to the text and the extra space goes to
the left/right/etc. of the icon.