Starting up a test suite for the Element API
We are generating a lot of reliance on wasAdded and wasRemoved, which are subtle and not quite up
to major use cases yet. It will never be awt, but we can perhaps have some middle ground.

In any case, this will help pinpoint the guarantees tripleplay customers need. As a case in point,
both trasfer tests will fail without the next commit.