Set the underline just above the bottom of descenders.
I looked around a bit for help here. This is the best I came up with:

http://graphicdesign.stackexchange.com/questions/11544/should-underline-clash-with-text-descenders

Sounds like the web "standard" has been to put it just below the baseline, but that looks crowded,
and is only like it because that's how hyperlinks always looked. The slightly better (although
modern typography apparently discourages use of underlines as much as passible) alternative is to
put it just above the bottom of descenders, as it interfers with the text less, but still doesn't
require extra height in your leading due to putting it all the way below all text.