Added AnimGroup for composing groups of animations.
You can use it to package up one or more parallel chains of animations and have
those queued up for actual animation at some later time. Or you can use it to
start multiple animations in parallel as part of an existing sequence of
animations. Or you can use it to make toast.

Needs testing/demonstrating. That's coming next.