Animation time is in milliseconds.
We discovered, due to the new changes, that we were being inconsistent in Yotab. I had used seceonds
in one case, but Jamie had used milliseconds in another and because they were being updated
separately, it was fine. Now that we're both hooking into the same global animator, it was a
problem.

AnimDemo agrees with Jamie (that it should be milliseconds - unless it really is meant to run over
1000 seconds), so I'm updating the docs and default here, assuming that that was the intention.