Attempted fix for garbage pixels on text
Having the glyph use fractional size was possibly causing issues further down the line. This is
easily observable in our game but not in any of the demo screens and I haven't had the time to
attempt a test case. This change appears to have fixed the problem in cursory testing, but then
again 0 is a common number :)