First pass at native textfield support for Java.
There are still lots of TODOs (real transparency, fonts, input modes,
etc), but it does the job well enough for now. Annoyingly this requires
that JavaTPPlatform manage it's own JFrame, which is sure to have
implications (JavaPlatform.setTitle not working, for one). This can be
helped a bit with changes up in PlayN later on.