Fiddling with virtual keyboard return key label.
This is going to be hard to support on both Android and iOS. This implementation is fine for iOS,
but it's unclear how well something like this can be done on Android. Looks like the user of an
Android device can actually go in and add whatever crazy ass keyboard he wants, so who knows how
much control applications have over something like this.