Initial bits for multi-line native text fields and java impl
In order to do this, ios and java are both going to need to reallocate their undelying fields.
Rather than relying on the somewhat broken approach in place for JPasswordField, extract the
concept of a mode into the interface and let Field structure things more sensibly. The native
field is only responsible for knowing when its implemenation needs to be reallocated based on
a mode request.