Support multi-line native text fields on ios
I needed to switch to UITextView to get multiple lines, but didn't want to commit to using that for
everything, so refreshMode does the swap over if Field.MULTILINE is set. This is still kind of
experimental but was working fine for my use case.

There isn't much new code, most of the changes are simply to isolate the usages of the various
UIKit pieces: UIView for most things; traits for the text attributes; and abstract methods
to call something specific to the leaf field type. The handler also needed to be aware of the
two types.