Improve menu positioning
The main missing piece here was that menus need to try and avoid popping up right over where the
pointer is. We assume this is pop.position for now, which it normally is (and when it isn't it is
already outside the bounds). Also decided to move the positioning code into RootLayout for a
cleaner interface between the two.