Clamp to make sure we don't have overlapping segments.
This would be kind of weird for a carefully design Scale9 source graphic, but in the default case
where a graphic is sliced up into 3 equal parts and then applied to a provided size, it generally
will look better to cut things short than overlap them.

Or at least, it definitely looks better in a few particular cases in yotab.