Remove the view from the list first.
This is needed so that if you're using the virtual keyboard handler to modify the keyboard hiding
behavior, we don't cause two finishedEditing signal dispatches (which also caused a hard crash).
I don't *think* there are any additional side effects creeping in from this change, but this stuff
is getting lousy with corner cases, so there could be something I'm forgetting.