When the view rotates, kill the keyboard if it's up.
Right now, the game view ends up in a bad way if you rotate the game with the keyboard up (and
therefore the game view has been slid out of the way). Ideally, the view would transform
properly for the new view rotation, but that wasn't super easy after a little fiddling and I need
to move on to other bugs, so for now just kill the keyboard and let the game view settle. I can't
imagine it will be a common case for people to want to rotate the device and keep typing away...

Although as I say that I realize that'll probably be true for puzzle pirates, but in a lot of
phone apps I frequently have it in portrait, pop the keyboard, and then rotate to lanscape to
continue typing. So we should probably support that behavior as a general feature of TP.