Fix double menu pop bug
This has been around for a while and not fixed because it's usually a logical error to popup a 2nd
menu before the previous one is finished animating. However, very fast users can cause a crash
by tapping quickly on a button that pops a menu. Fix by invoking the cleanup code later.