Fix (another) problem with menus and overlays
Because menus deselect their items on close, layout occurs a moment after the item trigger, which
causes a black box to appear on the overlay.

Also added a comment regarding how this ought to get fixed. I have a local commit to attempt to
support ImageOverlay, but it's not very far along.