New more abstract element parent: Container.
Elements is a Container, as will be another forthcoming container. This is
pretty internal, so hopefully this isn't going to break a bunch of clients.

I'm adding one new container, but this opens the door to other kinds of
containers which can now expose just the API they need, rather than extending
Elements and exposing an "I'm a list of elements" API in addition to whatever
sensible API they might have. (I'm looking at you Tabs.)