A new container type: Box.
It contains zero or one elements and passes things pretty much straight through
to its contents. This is useful if you have one "slot" in your UI that you want
to fill with, say, a "Loading..." label and then the real UI later, but you
don't want to have to pluck the Loading label out of its group and insert your
contents into the same index and blah blah.