We can now use the stock PlayN shader, pretty much.
Fixes #36.

Now that it supports tinting (more correctly than we did), it does everything
we need it to do. I had to rearrange it to ensure that all stable (same for all
verts in a quad) attributes were contiguous. Now that it is, we can splat all
of our per-particle data in a single copy, then add our vert+tex coords.

This fixes Evgeni's issue where tinting was causing weird edge artifacts.