Allow Behavior to apply to Element
This already works, it's just changing the type. This is really nice because a Container type
can now easily be made clickable using Behavior.

We may want to call enableInteraction in the Behavior constructor or at least expose some way
of doing it here so that there aren't those two calls to set HIT flags strewn around everywhere.