Added hacky mechanism to skip transition frames.
As the wordy comment indicates, this is a simple way to "fix" the jerkiness
caused by a screen loading up a bunch of images in wasAdded(). The loading is
not long enough to merit making a big deal out of it with a loading screen or
something, like it's 100ms, but 100ms causes a big jerk when you're normally
expecting animation frames to arrive every 15ms.

A "proper" solution is to load images asynchronously and then only start
rendering once they're loaded (though then you still have the delay caused by
uploading textures to the GPU on the first frame), but that complicates the
shit out of client code with a bunch of image callbacks and blah blah blah, and
Java provides no good programming language constructs to keep that from
overwhelming your otherwise simple UI initialization code with boilerplate and
callback spaghetti.