Alternative fix for clicks on text fields not working.
The problem with making the canvas not focusable is that it totally breaks PlayN keyboard. Oops!

So instead, manually detect when the canvas is eating a click and issue a focus request on the hit
component.

I also decided to bottleneck the overlays through the platform rather than exposing the JFrame,
which was sort of on the cards anyway, though not technically needed for this.