Sort out popup position configuration.
This was confusing the concept of trigger point and menu origin. There are some common scenarios,
but you basically want to be able to configure them independently. Now that can be done through
styles. The old "at" methods are deprecated but still supported by manipluating a couple of lazily
created members.

To convert to the new system, addStyles on the trigger and menu and be sure and pass the pointer
event to the Pop constructor.