Update the native text on the AWT thread.
This was causing a weird side effect. Highlighting the text in a field and typing a new letter
would place the insertion point at the beginning of the field (about 90% of the time, sometimes
it didn't). So if you typed two or more keys, you'd rarely get what you want.

Maybe this will shore us up for a while, but I'm quite worried about the "dual event pump"
situation with AWT and playn fighting. There are other issues like this one, and probably some
crash bugs lurking, waiting for just the right use case.