Refactor to work with SWT's immutable control style.
Instead of enumerating the kinds of fields that might cause reallocation, just assume reallocation
can occur on any style change. This will work with SWT, makes the ios and "java" versions more
clear, and moves the platform-specific logic of when the mode should really change down into the
paltform, where it belongs. Tada!