Revamped text rendering based on NWO.
TextConfig is out, and TextStyle and StyledText are in. This also lays the
ground work for rendering "rich" text (once we get a smidgen more support from
PlayN). TextStyle will define the style of a span of text, and StyledText will
have one more create() method which takes an array of strings and styles which
will be wrapped into a specified wrap width.