Make setPosition paint with a dt instead of setting _position
Previously, it would diddle _position itself and then paint with
a zero dt. Everything worked fine for simple movies, but if you had
a Movie which had other Movies on it, they wouldn't advance.

This is still a little bit weird for Movies that contain other Movies
with different durations; as you loop, the two fall in and out of sync,
but that's always been the case. If you're doing complicated movies like
that and then explicitly reaching in and modifying the position yourself,
caveat emptor.

Possibly the answer here is to let you setDuration with negative values?
Since it WILL let you set an arbitrarily overly-large value right now and
loop through N times to reach it. But at that point, maybe you should just
be calling paint() with your calculated dt yourself. I haven't fully thought
it through, so that's left as an excercise for the reader when he or she
decides to go into wacky town like that.