Vastly simplified animation completion.
Firstly, the logic was scattered all over the damned place and the copies
weren't consistent. AnimGroup was checking canceled, but Animation wasn't. The
duplication is now gone and the logic is necessarily consistent because it only
exists once.

I changed the canceled check to cause things to blow up. Either you cancel()
your animations or you complete() them. You don't mix and match. There's no way
we can do anything reasonable in such circumstances, so we disallow it.

The makeComplete() (nee complete()) method is now documented as being
idempotent so that we can now simply go through the whole animation chain and
mark every animation is complete whether or not it has been previously
processed.

If a layer is in an animation chain multiple times, like:

  moveTo(layerA, pointA).then.moveTo(layerA, pointB)

and the first animation has already completed before complete() is called, then
the first animation will move layerA to pointA during the completion process,
but the second animation will move layerA to pointB immediately afterwards, so
aside from a miniscule amount of extra effort (possibly offset by vastly
simpler completion logic) there will be no visible funny business. Everything
will end up where it should be were the animation chain allowed to complete
normally.

It is possible that a totally separate animation chain was queued up to
manipulate a layer after the first one completed, but then that first one went
on to do other things and was complete()d later, thereby stepping on the toes
of the second spawned animation chain. But arguably you should be completing
all animation chains in that case and completing them in the order that they
were spawned, so you should still be OK. Though you really shouldn't be doing
such crazy shit in the first place.

I also removed Repeat _isCompleted tracking because it didn't make any sense.
At first I thought animation chains were intended to be reused, but that's not
possible. complete() also cancel()s the animations, and once an animation is
cancel()ed, it can never be reused. Moreover, once an animation is cancel()ed,
next() will never be called on it. So there are no circumstances in which
Repeat would need to return something special from next() after it had been
complete()d.